Revert "chore: refactor some symbols out of table/mod.rs into their own files"
This reverts commit 7392171344f0cbd3e05342ee835025360fc66671.
