+++ crates/core/src/table/mod.rs
@@ -1,5 +1,6 @@
 //! Delta Table read and write implementation
 
+use std::any::Any;
 use std::cmp::{min, Ordering};
 use std::collections::HashMap;
 use std::fmt;
@@ -32,13 +33,180 @@ pub mod config;
 pub mod state;
 pub mod state_arrow;
 
-mod checkpoint;
-mod columns;
+/// Metadata for a checkpoint file
+#[derive(Serialize, Deserialize, Debug, Default, Clone, Copy)]
+#[serde(rename_all = "camelCase")]
+pub struct CheckPoint {
+    /// Delta table version
+    pub(crate) version: i64, // 20 digits decimals
+    /// The number of actions that are stored in the checkpoint.
+    pub(crate) size: i64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    /// The number of fragments if the last checkpoint was written in multiple parts. This field is optional.
+    pub(crate) parts: Option<u32>, // 10 digits decimals
+    #[serde(skip_serializing_if = "Option::is_none")]
+    /// The number of bytes of the checkpoint. This field is optional.
+    pub(crate) size_in_bytes: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    /// The number of AddFile actions in the checkpoint. This field is optional.
+    pub(crate) num_of_add_files: Option<i64>,
+}
 
-// Re-exposing for backwards compatibility
-pub use checkpoint::*;
-// Re-exposing for backwards compatibility
-pub use columns::*;
+#[derive(Default)]
+/// Builder for CheckPoint
+pub struct CheckPointBuilder {
+    /// Delta table version
+    pub(crate) version: i64, // 20 digits decimals
+    /// The number of actions that are stored in the checkpoint.
+    pub(crate) size: i64,
+    /// The number of fragments if the last checkpoint was written in multiple parts. This field is optional.
+    pub(crate) parts: Option<u32>, // 10 digits decimals
+    /// The number of bytes of the checkpoint. This field is optional.
+    pub(crate) size_in_bytes: Option<i64>,
+    /// The number of AddFile actions in the checkpoint. This field is optional.
+    pub(crate) num_of_add_files: Option<i64>,
+}
+
+impl CheckPointBuilder {
+    /// Creates a new [`CheckPointBuilder`] instance with the provided `version` and `size`.
+    /// Size is the total number of actions in the checkpoint. See size_in_bytes for total size in bytes.
+    pub fn new(version: i64, size: i64) -> Self {
+        CheckPointBuilder {
+            version,
+            size,
+            parts: None,
+            size_in_bytes: None,
+            num_of_add_files: None,
+        }
+    }
+
+    /// The number of fragments if the last checkpoint was written in multiple parts. This field is optional.
+    pub fn with_parts(mut self, parts: u32) -> Self {
+        self.parts = Some(parts);
+        self
+    }
+
+    /// The number of bytes of the checkpoint. This field is optional.
+    pub fn with_size_in_bytes(mut self, size_in_bytes: i64) -> Self {
+        self.size_in_bytes = Some(size_in_bytes);
+        self
+    }
+
+    /// The number of AddFile actions in the checkpoint. This field is optional.
+    pub fn with_num_of_add_files(mut self, num_of_add_files: i64) -> Self {
+        self.num_of_add_files = Some(num_of_add_files);
+        self
+    }
+
+    /// Build the final [`CheckPoint`] struct.
+    pub fn build(self) -> CheckPoint {
+        CheckPoint {
+            version: self.version,
+            size: self.size,
+            parts: self.parts,
+            size_in_bytes: self.size_in_bytes,
+            num_of_add_files: self.num_of_add_files,
+        }
+    }
+}
+
+impl CheckPoint {
+    /// Creates a new checkpoint from the given parameters.
+    pub fn new(version: i64, size: i64, parts: Option<u32>) -> Self {
+        Self {
+            version,
+            size,
+            parts: parts.or(None),
+            size_in_bytes: None,
+            num_of_add_files: None,
+        }
+    }
+}
+
+impl PartialEq for CheckPoint {
+    fn eq(&self, other: &Self) -> bool {
+        self.version == other.version
+    }
+}
+
+impl Eq for CheckPoint {}
+
+/// A constraint in a check constraint
+#[derive(Eq, PartialEq, Debug, Default, Clone)]
+pub struct Constraint {
+    /// The full path to the field.
+    pub name: String,
+    /// The SQL string that must always evaluate to true.
+    pub expr: String,
+}
+
+impl Constraint {
+    /// Create a new invariant
+    pub fn new(field_name: &str, invariant_sql: &str) -> Self {
+        Self {
+            name: field_name.to_string(),
+            expr: invariant_sql.to_string(),
+        }
+    }
+}
+
+impl DataCheck for Constraint {
+    fn get_name(&self) -> &str {
+        &self.name
+    }
+
+    fn get_expression(&self) -> &str {
+        &self.expr
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
+
+/// A generated column
+#[derive(Eq, PartialEq, Debug, Clone)]
+pub struct GeneratedColumn {
+    /// The full path to the field.
+    pub name: String,
+    /// The SQL string that generate the column value.
+    pub generation_expr: String,
+    /// The SQL string that must always evaluate to true.
+    pub validation_expr: String,
+    /// Data Type
+    pub data_type: DataType,
+}
+
+impl GeneratedColumn {
+    /// Create a new invariant
+    pub fn new(field_name: &str, sql_generation: &str, data_type: &DataType) -> Self {
+        Self {
+            name: field_name.to_string(),
+            generation_expr: sql_generation.to_string(),
+            validation_expr: format!("{field_name} = {sql_generation} OR ({field_name} IS NULL AND {sql_generation} IS NULL)"),
+            // validation_expr: format!("{field_name} <=> {sql_generation}"), // update to
+            data_type: data_type.clone()
+        }
+    }
+
+    pub fn get_generation_expression(&self) -> &str {
+        &self.generation_expr
+    }
+}
+
+impl DataCheck for GeneratedColumn {
+    fn get_name(&self) -> &str {
+        &self.name
+    }
+
+    fn get_expression(&self) -> &str {
+        &self.validation_expr
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
 
 /// Return partition fields along with their data type from the current schema.
 pub(crate) fn get_partition_col_data_types<'a>(
@@ -70,8 +238,6 @@ pub struct DeltaTable {
     pub state: Option<DeltaTableState>,
     /// the load options used during load
     pub config: DeltaTableConfig,
-    /// Kernel engine sitting under the [DeltaTable]
-    //engine: Arc<dyn delta_kernel::Engine>,
     /// log store
     pub(crate) log_store: LogStoreRef,
 }
@@ -488,5 +654,22 @@ mod tests {
             .await
             .unwrap();
         (dt, tmp_dir)
+    }
+
+    #[test]
+    fn checkpoint_should_serialize_in_camel_case() {
+        let checkpoint = CheckPoint {
+            version: 1,
+            size: 1,
+            parts: None,
+            size_in_bytes: Some(1),
+            num_of_add_files: Some(1),
+        };
+
+        let checkpoint_json_serialized =
+            serde_json::to_string(&checkpoint).expect("could not serialize to json");
+
+        assert!(checkpoint_json_serialized.contains("sizeInBytes"));
+        assert!(checkpoint_json_serialized.contains("numOfAddFiles"));
     }
 }
